does some initial code cleanup, addressing [clean up code management for dev example #40] and [Move utility functions into utility.js #23]. note that this changes the local development process - see README.md
