+++ dev-example/predevals-options.json
@@ -1,1 +1,103 @@
-{"targets":[{"target_id":"wk inc flu hosp","metrics":["wis_scaled_relative_skill","wis","ae_median_scaled_relative_skill","ae_median","interval_coverage_50","interval_coverage_95"],"relative_metrics":["wis","ae_median"],"baseline":"FluSight-baseline","disaggregate_by":["location","reference_date","horizon","target_end_date"]}],"eval_sets":[{"eval_set_name":"Full season","round_filters":{"min":"2024-11-30"}},{"eval_set_name":"Last 4 weeks","round_filters":{"min":"2024-11-30","n_last":4}}],"task_id_text":{"location":{"US":"United States","01":"Alabama","02":"Alaska","04":"Arizona","05":"Arkansas","06":"California","08":"Colorado","09":"Connecticut","10":"Delaware","11":"District of Columbia","12":"Florida","13":"Georgia","15":"Hawaii","16":"Idaho","17":"Illinois","18":"Indiana","19":"Iowa","20":"Kansas","21":"Kentucky","22":"Louisiana","23":"Maine","24":"Maryland","25":"Massachusetts","26":"Michigan","27":"Minnesota","28":"Mississippi","29":"Missouri","30":"Montana","31":"Nebraska","32":"Nevada","33":"New Hampshire","34":"New Jersey","35":"New Mexico","36":"New York","37":"North Carolina","38":"North Dakota","39":"Ohio","40":"Oklahoma","41":"Oregon","42":"Pennsylvania","44":"Rhode Island","45":"South Carolina","46":"South Dakota","47":"Tennessee","48":"Texas","49":"Utah","50":"Vermont","51":"Virginia","53":"Washington","54":"West Virginia","55":"Wisconsin","56":"Wyoming","60":"American Samoa","66":"Guam","69":"Northern Mariana Islands","72":"Puerto Rico","74":"U.S. Minor Outlying Islands","78":"Virgin Islands"}}}
+{
+  "targets": [
+    {
+      "target_id": "wk inc flu hosp",
+      "metrics": [
+        "wis_scaled_relative_skill",
+        "wis",
+        "ae_median_scaled_relative_skill",
+        "ae_median",
+        "interval_coverage_50",
+        "interval_coverage_95"
+      ],
+      "relative_metrics": [
+        "wis",
+        "ae_median"
+      ],
+      "baseline": "FluSight-baseline",
+      "disaggregate_by": [
+        "location",
+        "reference_date",
+        "horizon",
+        "target_end_date"
+      ]
+    }
+  ],
+  "eval_sets": [
+    {
+      "eval_set_name": "Full season",
+      "round_filters": {
+        "min": "2024-11-30"
+      }
+    },
+    {
+      "eval_set_name": "Last 4 weeks",
+      "round_filters": {
+        "min": "2024-11-30",
+        "n_last": 4
+      }
+    }
+  ],
+  "task_id_text": {
+    "location": {
+      "US": "United States",
+      "01": "Alabama",
+      "02": "Alaska",
+      "04": "Arizona",
+      "05": "Arkansas",
+      "06": "California",
+      "08": "Colorado",
+      "09": "Connecticut",
+      "10": "Delaware",
+      "11": "District of Columbia",
+      "12": "Florida",
+      "13": "Georgia",
+      "15": "Hawaii",
+      "16": "Idaho",
+      "17": "Illinois",
+      "18": "Indiana",
+      "19": "Iowa",
+      "20": "Kansas",
+      "21": "Kentucky",
+      "22": "Louisiana",
+      "23": "Maine",
+      "24": "Maryland",
+      "25": "Massachusetts",
+      "26": "Michigan",
+      "27": "Minnesota",
+      "28": "Mississippi",
+      "29": "Missouri",
+      "30": "Montana",
+      "31": "Nebraska",
+      "32": "Nevada",
+      "33": "New Hampshire",
+      "34": "New Jersey",
+      "35": "New Mexico",
+      "36": "New York",
+      "37": "North Carolina",
+      "38": "North Dakota",
+      "39": "Ohio",
+      "40": "Oklahoma",
+      "41": "Oregon",
+      "42": "Pennsylvania",
+      "44": "Rhode Island",
+      "45": "South Carolina",
+      "46": "South Dakota",
+      "47": "Tennessee",
+      "48": "Texas",
+      "49": "Utah",
+      "50": "Vermont",
+      "51": "Virginia",
+      "53": "Washington",
+      "54": "West Virginia",
+      "55": "Wisconsin",
+      "56": "Wyoming",
+      "60": "American Samoa",
+      "66": "Guam",
+      "69": "Northern Mariana Islands",
+      "72": "Puerto Rico",
+      "74": "U.S. Minor Outlying Islands",
+      "78": "Virgin Islands"
+    }
+  }
+}
