+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ajv-play",
+  "name": "predevals",
   "version": "1.0.0",
   "type": "module",
   "devDependencies": {
